+++ src/Domain/Rotation/RotationFactory.ts
