+++ src/Shared/Errors/IncompatibleMatricesError.ts
