+++ src/Shared/Errors/MatricialProductError.ts
