+++ src/Shared/Errors/MatrixError.ts
