+++ src/Shared/Errors/MatrixSizeError.ts
